add register page add new class TextItems to use with text widgets which should be updated on lang change etc minor modifications (including compile fix)

+++ database.cpp
@@ -215,7 +215,7 @@ int Database::ExecuteQuery()
     }
 
     #ifdef DEBUG
-    printf("\n\nExecuteQuery(): test6: rows.size(): %i\n", rows.size());
+    printf("\n\nExecuteQuery(): test6: rows.size(): %i\n", (int)rows.size());
     #endif
 
     return rows.size();
